Refactored the Code
