+++ src/classes/chains/Ethereum/EthereumCollection.ts
