+++ src/classes/chains/Ethereum/EthereumContract.ts
