+++ src/classes/fileStorage/NFTstorage.ts
