+++ src/classes/fileStorage/Storj.ts
