+++ tsconfig.json
@@ -67,5 +67,5 @@
 		// "skipLibCheck": true,                  /* Skip type checking of declaration files. */
 		// "forceConsistentCasingInFileNames": true   /* Disallow inconsistently-cased references to the same file. */
 	},
-	"include": ["./src/**/*"]
+	"include": ["./src/**/*", "examples/chains/Ethereum/contracts/EthereumCollection.ts", "examples/chains/Ethereum/contracts/EthereumContract.ts"]
 }
